Front: Move 'SongWithRelation' model to its own file To avoid circular import
